Metadata override feature
The key purpose of this feature is to enrich metadata with additional
properties (keys), and to replace values of existing properties (keys)
with other values. The "amend" use case (add values to a particular key)
is only supported vai the proxy of replacing all values with a copy of
the original values that has new items appended (because @mih thinks
that this case is not common in typical metadata enrichment scenarios).

The current implementation allows for a unchanged TSV (author-provided)
to be combined with an evolving override (curator-provided) to inject
additional properties (e.g., `@type`, or `@id`) in the metadata record
without altering the structure/content of the author-provided dcouments,
thereby avoiding any friction or incompatibilities with the workflows
or processes that yielded them in the first place.

This makes it easier to feed back corrections to the original authors
(or ask for them), and does not require any party to adjust to a
different workflow.