Make `load(mode=)` less about recursion only
With the addition of a JSON-LD context, it made sense to me to
distinguish three modes:

- `jsonld` (the real deal, recursion and contexting)
- `json` (no semantics, but will recursion)
- `single` (no recursion, no context)

No recursion, but contexting does not seem to make sense. If there are
no imports this is the same as `jsonld`. With imports, the context
will make little sense, because the structure is incomplete.