Support record-wide (base) context definition
The context can be declared at `<prefix>.ctx.jsonld`.

It can be used to host *all* context information applicable to any
sheet. Alternatively, it can be used as a base context, possibly
with just the used ontologies defined, that is then amended, per sheet,
with the relevant terms, or property vocabulary definitions.

The `demorecord` is adjusted according to the latter approach.

Closes #70