Support 'single-record-per-directory' format
In preparation for
https://github.com/psychoinformatics-de/datalad-tabby/issues/79 and
despite the conclusion in
https://github.com/psychoinformatics-de/datalad-tabby/issues/50 this
change adds support for a simplified set of files that form a tabby
record.

The only thing that is simplified is that the common prefix is removed
from all filename. The demo record is not also included in this format.

This layout is what we would like put into a ZIP file container.

The prefix continues to exist (this was the main concern in #50), but is
now the name of the parent directory.

In https://github.com/psychoinformatics-de/datalad-tabby/issues/55
this simplifies the setup for the self-description of a dataset. All
files could go into `.datalad/tabby/self/` and have short names like:

- `dataset.tsv`
- `dataset.override.json`
- ...

There is no particular additional markup necessary to distinguish
single-item-dir format from the prefixed-layout. The absence of an
underscore char, is evidence enough.

Closes #50 (for real)