tabby extractor injects minimal ID and type info
This established the absolute minimum necessary to distinguish
summary-level and version-level dataset descriptors.

The main metadata record is always considered to be a version-level
description. In addition, a summary level description is linked
via `dcterms:isVersionOf`. The summary-level descriptor itself
is bare-bone and only declares `hasVersion` with a backlink to the
version-level description.

Ping https://github.com/psychoinformatics-de/datalad-tabby/issues/76