tabby extractor also reports `dscollection` records
They are reported as-is, no @id fiddling or anything.  In principle this
need not be limited to a _dataset_ collection.  However, we somehow need
to "find" the root sheets to load the record. Going for `*_dataset.tsv`
seems simple.

Would be useful to have a utility to find records in a "dump".