Support JSON-only tabby; and many-table object templating
This respective tests are included.

In addition to injecting any data in `many` imports via a JSON array
deposit, JSON-object templating is now also supported, by putting
a JSON-object into the JSON file corresponding to a many table.
The information in such an object becomes the template for any object
in the many table, and the information from the table amends this
object.

Closes