Improve `size[bytes]` definition
Previously the interpretation was subject to some kind of postprocessing
of a potentially existing unit in the string value https://schema.org/contentSize

We switch to https://www.semanticdesktop.org/ontologies/2007/03/22/nfo/#fileSize
now, which is much more precise:

    The size of the file in bytes. For compressed files it means the size of
    the packed file, not of the contents.

Moreover, we declare the data type to be integer. This is useful,
because when read from a TSV file, the native JSON dtype will be `str`,
not `int`.

Closes #103