Trim empty rows when converting xlsx to tsv
With this change, excel to tabby conversion will preserve empty lines
in the middle of the file (maybe a visual separation of sections in a
many-objects file), but truncate empty lines at the end (maybe excel
artefact). This requires double iteration over the rows (first to find
where data ends, then to export), but it seems inexpensive.

This should help situations when excel (or calc) xlsx file preserves
blank lines.

One test data file (tsv) used to test round-tripping is altered to
remove empty lines at the end. So in the end we no longer guarantee
round-tripping these empty lines, but I feel this was a non-feature.